backend/requirements.txt
# 基础架构
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-ulid==3.0.0
python-multipart==0.0.20
qrcode==8.0
pandas==2.2.3
pillow==11.1.0
# 定时任务, 清理二维码缓存
apscheduler==3.11.0

# 其他配置
pydantic-settings==2.7.1
openpyxl==3.1.5
python-dotenv==1.0.1